requirements.txt
dbt-snowflake==1.10.2
pandas==2.3.2
plotly==6.3.0
python-dotenv==1.1.1
requests==2.32.5
snowflake-connector-python==3.17.3
streamlit==1.49.1